+++ nixos/configuration.nix
@@ -21,6 +21,7 @@
     ./modules/schedule.nix
     ./modules/tripwire.nix
     ./modules/egress.nix
+    ./modules/governor.nix
   ];
 
   # The NixOS release this configuration was written against. Do not change this